[FIX] web_editor: prevent zws in url on paste
Paste a url in the editor from the URL bar of the browser. If you then
select it all and paste another URL in its stead, the zws inserted by
the sanitizer when deleting the selection ends up in the URL.
This is because the santizer moves downward into the DOM tree, therefore
change the link URL before removing the zws in its text child node.
To make sure this can't ever happen, we now remove zws from URLs when
parsing them from the link's text content.

task-3054233