[FIX] l10n_it_edi: ImportoTotaleDocumento should be in EUR
A client complains that while the "Divisa" (currency) is EUR
then the total aside of that is in DDK (Danish Crown).
The specs from Agenzia delle Entrate (Italian Tax Agency)
are unclear, they say that we can retain the total in DDK
but we must specify it "somewhere in a descriptive field".
Then it makes sense and it's more linear to just change
the total to EUR too.

Specs: https://www.agenziaentrate.gov.it/portale/web/guest/schede/comunicazioni/fatture-e-corrispettivi/faq-fe/risposte-alle-domande-piu-frequenti-categoria/fatture-verso-e-da-soggetti-stranieri-transfrontaliere#collapse-1750220