[FIX] web: assign appropriate sequence value to new records
Before this commit, when a new record with a sequence field was created in an
x2many field rendered in a list view, the initial sequence value was it's
default value. This results in weird behavior when moving the new record to a
new position in the list. Instead, with this commit, the initial sequence value
reflects the record's initial position in the list.

Note: in the original opw ticket, the problem occurred when adding new lines to
a sale order based on a sale order template. From the bug report it seemed the
problem only occurred when the sale order template did not contain any section
lines. However, this is not really what happened:

1) A sale order template is created with only product lines. The lines of the
   template itself are added in a list view renderer, so they all have their
   initial default sequence values, which happens to be 10.
2) A sale order is created based on this template. The sequence values are
   copied from the template.
3) A product line is now added in the list view, and subsequently moved to a
   new position. The BasicRenderer detects in the _moveRecord that some lines
   have the same sequence value as the previous line, and now decides that
   reordering all the lines is appropriate. As a consequence the move works as
   expected.
4) Back to the sale order template: the bug report now describes adding two
   section lines. These are both also moved to new positions. For the same
   reasons as before the BasicRenderer decides to reorder all the lines. As a
   consequence the template lines no longer all have the same sequence number.
5) This updated template is now used to create another sale order. However
   the lines in this sale order have appropriate sequence values 10, 11, 12 and
   so on.
6) Again a product is added in the list view, this lines has the default
   sequence value which is 10. When moving this line, the BasicRenderer does
   not decide to reorder all lines. Instead, only the lines contained between
   the new and the old position of the moved line are reordered. New sequence
   values are written on these lines, starting from the sequence value of the
   newly added line.
7) As a result, the upper part of the lines keeps their sequence values 10, 11,
   12 and so on. The lower, reordered, part gets new sequence values which,
   incidentally, also start with 10, 11, 12 and so on. In the end both parts
   end up interleaved.

As is clear from looking at these steps in more detail, the presence of the
section lines played no part in the bug.

opw-2916207