[ADD] web_editor: Adding new toggle list
This commit is adding a new type of list inside of the editor
called the `togglelist`.
This list is composed of items and divs linked to those items.

The idea is that if you click on the toggle the text inside the
div is hidden or shown depending on the state of the toggle.
If the arrow of the toggle points down the text is shown, otherwise
if the arrow points right the text is hidden.

task-2796594