[FIX] mail: make auto-delete more consistent
'auto-delete' can be set in many ways as there are multiple checks
inside the composer on these different inputs, sometimes not equivalent.
This consolidates the condition for auto-delete into a single function.

Additionally mailings were not consistently setting auto-delete
even if the keep_archives option was off. The auto-delete value is now
set in the composer explicitly.