[FIX] stock: fix picking_type creation without WH
Avoid raising an error if `warehouse_id` isn't in the vals.

Steps to reproduce :
Inventory -> Configuration -> Operation Types -> Create a new Operation
Type.

The `warehouse_id` can't be manually set until the multi-warehouse is
activated. So until it is activated, the warehouse is set to the default
warehouse of the company