[IMP] im_livechat, mail, test_mail: allows unfollowing record from inbox
Allow users to unfollow easily document from the inbox for which he/she doesn't
want to be notified anymore.

When hovering on a follow up message with a related document followed by the
user, the interface displays an action to unfollow it that allows the user to
remove himself as follower of the document. As a result of performing that
action a toast is displayed as a feedback.

Technical note:
The unfollow link is displayed on a message when its property can_unfollow is
true; that computed property is true when the related focument is followed by
the partner.

To know if the partner still follow the document, we have extended the
mechanism that load partially the related thread of the message when loading
initial messages by adding also current partner as follower if the partner
follows the related document.

Necessary data being setup, the unfollow action can then use standard mechanism
already developped to unfollow a document.

Nonetheless, to support unfollowing a document in the inbox no matter the
current company, we have modified message_unsubscribe in mail_thread to allow
internal user to unsubscribe themself without checking any rights. Indeed, some
document have record rule that prevents reading it if the user is not in the
right company (ex. crm.lead) and then was preventing the user to unfollow the
document using that method.

Note that the computed field follower_id in mail.notification model may not
always up to date as it doesn't depend on mail.followers.

Task-3061864