[ADD] web: add AdvancedSelect component
This commit adds a new AdvancedSelect component. This component makes it
possible to search for options for a select element. It can be used to
replace the jquery select2 element for most cases. Many options allow the
developer to custom the usage and look of the element for its own needs.

This commit also contains a Select component which allow advanced and native
html selection (although we are unsure if we should have a component for
a native select).

Tests have been added to make sure the behavior of this component is
working as expected, and let the core element's usage being asserted.

Co-authored-by: luvi <luvi@odoo.com>