[FIX] bus: pool error on trigger open event
Before this commit, failing to acquire a cursor for the `OPEN` event
would lead to a `PoolError` being raised.

When the websocket connection is properly opened, it is closed with
the `TRY_LATER` close code but since the connection is not yet opened,
let return a `ServiceUnavailable` answer to the client.