[IMP] mail, test_mail, website_slides: allows unfollowing record from email
Allow partner to unfollow a document from a follow up email of that document
through an unsubscribe URL in the email even if not connected.

It works for internal user for follow up on any document and for any partner on
follow up of document tagged as authorizing being unfollowed by any partner (
slide.channel and slide.slide).

Technical note:
When sending a notification by email, the recipients are splitted further
depending on being follower of the related document or not. If they are
followers a generic unfollow link is added in the mail.mail body_html. As a
mail.mail can be used to send mail to multiple recipients the link is refined
when sent to a specific recipient by setting the parameters (pids, token,
model, res_id). It is also when sending the mail that we check that the user
can have an unfollow link (user must be internal or the document is tagged as
authorizing to be unfollowed by any partner).

Rather than splitting recipients further into different mail.mail, we could
have added whether each recipients of a mail.mail is following the document or
not but we didn't find proper way to do that (info on relation between
mail.mail and partner ?).

An other possibility would have be to query follower when sending mail but that
would defeat the strategy to do the work upstream to be able after to send the
mail in batch easily.

Note also that we use special tags in the mail layout template
(unfollow_placeholder_start, unfollow_placeholder_end) for 2 purposes:
- to allow to remove that section for some recipient without rerendering the
template and avoid extra queries.
- to not include the unfollow link for partners that doesn't follow the
document (it is implicit, the group of recipient not following the document are
added in a mail.mail with body_html with that section removed)

Task-3061864