[ADD] sale: possible to add products from catalog
Before this commit to add products to a SO we needed to add one
product at a time and then reclick on the 'Add product' and repeat.
For large SO this become cumbersome fast.
After this commit and action 'Add from Catalog' will be added that
will allow to add products on batch from the catalog, making more
easy to manage sales orders.

We inherit from mail_thread to avoid changes to the sale order while
on draft. After discussion with the sm team and consideration this
seems to be the least intrusive way to do it.

See odoo/enterprise#37915

Task - 3062080