[IMP] mail: link messaging menu previews to threads
Before this commit, messaging menu previews were relying
only on 1st fetch of previews. There was no integration with
known threads in modelling. For example, new messages or new
pinned channels were not shown in messaging menu if both happen
after 1st opening of this menu.

This commit fixes the issue by integrating messaging menu previews
to modelling of threads, so that it shows data on threads, such
as the most recent message.