[FIX] mail: fix im_status crash on unknown partner
Before this commit, the im_status notification coming from the bus
would lead to a crash if the partner is not yet known. This is due
to the fact that the notification handler assumes the partner exists
thus writing on undefined.

This PR fixes this issue by inserting the partner instead of manually
udpating its im_status. Therefore, the partner is either created or
updated with the values received from the server.