[FIX] mail: fix attachment loading spinner when creating record
Before this commit, the attachment button would have shown the loading
spinner when creating a new record.

This was due to the fact that the loading state was based on the request
list passed to the `mail/thread/data` route. The spinner was removed
after receiving the answer.

In the case of non yet created record, the load method returns without
calling the server, thus preventing the load state to end.

This PR fixes this issue by setting the load state when we are sure that
the record exists.