[FIX] web_editor: prevent error an font-size change
Before this commit, changing the font size of an element that is
redundant to one of its ancestor could trigger a traceback. For fixing
that issue some code has to be refactored to handle the setFontSize case
in formatSelection. As a side effect of this refactoring, nested
font-size is now properly handled (see added tests).

task-3086301