[FIX] mail: selection when ref has been unmounted
Before this PR, the useSelection hook would access the element of its
ref without checking if the element was defined. This makes the chatter
crash when hidding the composer, and could also make a message in
editing crash for the same reason.

This PR fixes the issue by guarding ref.el calls since they make no
sense when the component is being unmounted.