[FIX] mail: thread model partial update
Since the introduction of the update method in the thread model,
updating a thread record with partial informations is broken.
Indeed, all non-present informations are set to undefined,
resulting in crashes.

For exemple, send an attachment in a message: the originThread
in the notificaiton only contains id and name values. This breaks
the thread record as the only keys left are id and name.

To solve this issue, non-present keys are defaulted to the values
already present on the record thus allowing partial updates.