[FIX] mail: starred/history reply to
Before this commit, the reply to action was available on starred/history
messages but was not handled correctly. On the master branch, this
functionnality is not available.

It costs nothing and makes no sense to disable this option for the two
other mailboxes.

This commit fixes the issue thus allowing to reply to messages in any
mailbox.