[FIX] base: should read PropertiesDefinition field from child
Steps to reproduce:
Create an article Parent A and set it as private
Create a child article, Child A
Invite a user on Child A so that they can read it
Log in as a user and try to open Child A from either a direct link or a ctrl+k

Before this commit:

It shows an access error regarding the parent article because of the 'Properties
Definition' field.

After this commit:

Without any access error user will be able to read that article, even if it
cannot be displayed in the menu because he can't read its parent.

Task-3062064