[IMP] crm, test_event_full, test_mail_full, test_mail: update query count
Query count adjustment.
Based on some test (not complete), the added query counts are due to the query
of user_follower_id in message_format of mail.message. It is necessary to
inform the web client which message is related to a document that the user
follows in order to display or not the unfollow link and to be able to
use the already developed method on the client side (which requires the
mail.followers id).

Task-3061864