[FIX] mail: fix channel suggestions
Before this PR, the processing of the channel suggestions
from the server caused a crash. This was due to the fact that
a Thread.insert was done without any model.

Also, channel mentions could not be posted since it was retrieving
the thread from its id instead of its local id.

This commit both the issues.