[IMP] web: adapt js for boolean in domain and no pre-processing
Following the change in the semantics of the domains, and the new
semantics of the views for `required`, `readonly` and `invisible`,
these receive domains which are not pre-processed. Previously
attributes such as `invisible="context.get('hide_me')"` were evaluated
on the python side before being sent to the web client. The javascript
is adapted to interpret `[true]` domains as always being true.

opw-2495504