[FIX] mail: mail.message._search() now makes less queries
Combine both queries made in _search() to retrieve accessible records.
This simply uses the API of the Query object to retrieve the data that
is necessary to restrict access to messages.

Move security check outside of _message_format() for performance.  The
call to check_access_rule() inside _message_format() was redundant in
many cases and generated more SQL queries than necessary.

Also for performance, accessing fields from records should not actually
check permission.  That's a bit freaky, but this reproduces the former
behavior of mail.message.

And finally, make method check_access_rule() check ir.rules, in order to
make it consistent with method _search().