[WIP] Project: refactor personal stage logic
In the module project, the model project.task.stage.personal was used to
make the link between user, task and personal stage (record of the model
project.task.type with no project_id set). This logic was difficult to
conciliate with the one used in todo and would have lead to too much
complexity in the code. It is therefore simplified to not used this
intermediate model anymore and use a logic similar to the one used in
note.

task-3085077