[FIX] hr_contract: fix date of contracts in demo data
Before this commit, when a new database with demo data are initialized
31 December and the unit tests are launched the day after, some tests
could fail if a contract running is expected for a certain resource.
Indeed, some contracts running could be expired because the dates of
those contracts are set on the year in which the demo data are added.
It means with the new year transition, the dates of some contracts
in demo data are inconsistent.

This commit reviews the dates of those contracts to be sure the contract
date will still be consistent and so the running contract will not
expire with the new year transition.