[FIX] stock_account: ensure product category configuration
Since [1] when a user sets the inventory valuation of a product
category to _Automated_, we try to ensure that the account
properties are defined (in both `create` and `write`). For instance:
https://github.com/odoo/odoo/blob/9dc7835cf62302ce05db822dcc1db9e5b577338f/addons/stock_account/models/product.py#L850-L854
As shown, we use the fields of the current company as fallback.
However, these fields are not mandatory. So, in case they are not
defined, the category will be incorrectly configured and the user might
not be aware of this. We should rather give the user some suggestions
(thanks to an onchange) and let him take the final decisions.

Also, on client side, these fields are mandatory:
https://github.com/odoo/odoo/blob/a5874da269fc5546f45a2aadd022eb9dbab166e7/addons/stock_account/views/product_views.xml#L22-L25
But there isn't anything on server side. We should add such a
constraint to ensure a correct configuration.

Last: [1] ignores `property_stock_journal` but this field follows
the same logic: it should be defined if the inventory valuation is
automated, and should not in the other case.

[1] 9dc7835cf62302ce05db822dcc1db9e5b577338f