[FIX] web_editor: prevent too deep range
The utility function `getDeepRange`, when correcting triple clicks,
sometimes ended up giving a range on a void element (eg, a <br>). This
is nonsense since by definition they have no children and therefore
can't have an offset. The consequences could be hard to predict. This is
one of them, which is fixed with this commit:

1. Open website builder
2. Add a "Cover" snippet
3. Select all contents of the heading
4. Insert a character
-> The browser appends an invisible <br> to the heading.
5. Select all contents of the heading, with the mouse going to the
   beginning of the next line
6. Insert a character
-> The style was lost because the <font> element in the heading was
removed.

task-3102841