[FIX] stock: don't allow modify cancelled move when unlock
When unlock a MO/PO, we can modify the done qty of a cancelled move.
However, this change won't be applied to the stock of the product since
the move is cancelled. To fix it, we don't allow to modify hte canceled
moves, users are advised to create a new move instead of changing the
cancelled ones.

Task-3116125