[IMP] web: x2many: disable form view modal opening
This commit disable by default the opening of the form view when
clicking on a record in a x2many list field.

Before when clicking on a record in a x2many list field:
- sometimes it goes into inline edition mode
- sometimes it opens the record in a modal form
- sometimes it does nothing

With this commit, we remove the "popup" option and the behavior becomes
more logical: it goes into inline edition if the field is editable,
otherwise nothing happens.

For specific cases, it is possible to activate this behaviour in the
tree arch with the `open_on_click="0"` attribute. This option was
previously named `no_open` but the double negation needed to open a
record (`no_open="0"`) is not clear enough.

If the x2many field renders a kanban, it stills open the records in a
modal by default because unlike the list, the kanban does not contain
all/most of the required information of the record.

task-id: 2964320