[IMP] hr,hr_contract: add demo for part time resource calendar
This commit adds a demo for part time calendar.
So the user gets a better idea of how contract, timesheet, and planning work.

We used employee 'Toni Jimenez' for the part time,
and set the tag 'Standard 20 hours/week' to identify easily.
Also, make the contract 'running' from the draft.

task-3128491