[REF] l10n_ke_edi_tremol: adapt invoice posting client action for owl
The client action that allows the user to post invoices to the Kenyan
fiscal device needs adapting to utilise the current JavaScript
framework.

This commit refactors the existing client action defined in
'l10n_ke_edi_tremol/static/src/js/post_send.js'.

The functionallity is mostly factored out of the action, and into a
service and a dialog that are utilised by the action.

The service is registered as 'l10n_ke_proxy' and handles all of the
logic related to posting an invoice to the device. It has a 'state'
attribute that represents the current progress in this posting process.

The dialog is called 'KEProxyDialog', and is passed to the dialog
service by the client action. It makes use of the state from the
'l10n_ke_proxy' service, which is uses to update the UI to inform the
user about the success/failure of posting the invoices.

The dialog uses the slots from the standard dialog, removing the close
'x' button from the header, and hiding 'ok' button from the footer
unless there has been an error (this gives the user time to read the
error message before closing the dialog). If all invoices are
succcessfully posted, then the dialog automatically triggers a reload.

Additionally, the function on the account.move model that decomposes the
results and writes them to the fields of the records has been renamed to
"l10n_ke_cu_responses" and its behaviour adapted to handle batched
responses.

The http_service.js is also modified to throw the appropriate 404 error
when the endpoint is not found.