[FIX] account: prevent posting autobalanced misc
The aim of this commit is to prevent the user to post a misc entry that
was balanced automatically. (move saved with debit != credit)

Context:
Since commit d8d47f9ff8554f4b39487fd2f13c153c7d6f958d we allow user to
creates unbalanced entries by automatically balancing them on a suspense
account. This isn't an issue but we also allow to post those entries.
As the flow is smooth and fast, a quick user could just post this
entries which could be problematic.
Resetting to draft could be an option but that wouldn't be convenient if some
lines would have been reconciled. (which could have side effects in case
of there would be a cash basis tax on the line)
It's also impossible in case of a journal that would be hashed.
It also means that the problem would be delayed until the end of year
accounting processes, when accountants are tied to dead-lines.

Before this commit:
An entry automatically balanced with a suspense account could be posted

After this commit:
The balancing line is created with a technical off balance account.
This account is deprecated and thus can't be directly selected by the
user.
When posting an entry with this technical account, a UserError is raised
asking the user to either delete the line and balance the move or to
change the account on the line.

community-PR:

task-id: 3129172