[FIX] web: crash when clicking deleted records in breadcrumbs
Here is a way to reproduce the problematic situation: with CRM
installed, go to Contacts, open one, clicks on its Opportunities
stat button, open one in form view, click on the "Customer" field
internal link to open again the contact in form view, delete the
record, and finally click on the first occurence of the contact
in the breadcrumbs.

Before this commit, it crashed because we ended up in an error
handler (formSaveErrorHandler) that doesn't deal with the fact
that originalError could be falsy. This is handled by the
legacyRejectPromiseHandler, but due to the low sequence of the
faulty handler, this one was executed after.

Besides the crash, the breadcrumb was weirdly updated (the contact
name was replaced by Contacts), and no other feedback was given to
the user.

This commit fixes the error handler sequences s.t. the crash no
longer occurs, prevents the breadcrumbs from being altered and
displays a notification indicating that the record does not exist.

Issue reported in the feedback pad after migrating odoo.com to 16.0.