[IMP] microsoft_outlook: allow to use the same token for sending / receiving
Purpose
=======
Allow to use the same token for outgoing mail server and for incoming
mail server. That way the user can login only once to configure both.

Make the code more consistent between google_gmail and microsoft_outlook
(use the same fields, same condition, same message, etc and simplify as
much as possible the mixins to remove code redundancy in child class).

Technical
=========
Create a new model to store the Outlook tokens. This model is now
responsible for refreshing the access token.

Because the token can be shared across multiple records, from different
models (`ir.mail_server` and `fetchmail.server`), we create a new token
when the email address changed, and we clean the table from unused
tokens in the autovacuum CRON.

Task-3061856