[IMP] account: tax update - better error message if country misconfiguration
We do not manage update of taxes if the fiscal country of the company is different
from the country on the chart_template. This setup doesn't make much sense but it's
not possible for us to deduce which one of those two countries should be used.
Thus we improved the catch of this misconfiguration and display a better error message
to invite the client to correct the setting himself.