[FIX] web: filter non-exportable fields server side
Since non exportable fields filtering could be done directly from the
server instead of doing it from the browser, this commit moves the
filtering logic to the server. Before this commit, it was still possible
to export those fields, since the attribute was not passed to the views
and the filter was only applied via the JS code. By moving the filtering
to the server, we remove the need to send the 'exportable' attribute
for the fields at each rpc call.

Now, that attribute is only read in the export controller to determine
if the current field must be send or not. This commit also prevent users
directly calling the export routes (without the web client, using direct
HTTP calls) to require non-exportable fields.

This change required some refactor of the export dialog, by renaming 'name'
to 'id', which simplify the code of the component.

The qunit test asserting the absence of non-exportable fields from the view
has been removed, since it was no longer coherent to test this from the JS
side. Instead, a python test has been added to verify that non-exportable
fields cannot be exported, even by calling the export route directly.