[FIX] web_editor: apply color only applies on first and last nodes
Before this commit:

If you write a log note, copy paste it, and then try to add color to it, the
color will only appear on the first and last node.

After this commit:

Now, when you color on text copied from a log note, it applies to all of it.

Task-3171392