[IMP] hr_contract: remove calendar mismatch warning
Before this commit, there was a warning popping either
on the employee or the employee's running contract page
whenever a mismatch between both was effective.
Since the employee calendar is in readonly if there is an
associated running contract and that the running contract
will automatically adapt the employee calendar to fit,
that warning is not really needed anymore.
This commit removes that warning

task-3177030