[IMP] stock, web: allow auto-printing + chained (report) actions
This commit adds two linked features:
1- the ability to chain (report) actions
2- the ability to auto-print delivery report at picking validation

Because feature 2 results in the possibility of 2 actions to be returned
at picking validation (due to pre-existing auto-open reception report at
validation option), feature 1 is required to support it. It is expected
that chaining will primarily occur with report actions since chaining of
the other action types does not currently have an applicable use case.
Since we also do not currently expect this chaining behavior in any
other (non-picking_type use case), for now we encapsulate the utility
method to chain the report actions  within stock [may change due to
future refactoring/meet future needs].

This is the first commit to support a series of potential
auto-printable reports.

Part of task: 3046178 - Operation Type