[FIX] account: bank statement ending balance default
Using the bank recon widget (enterprise), it is possible to create bank statements from the list view.
Two cases exist: 1. the line is saved, and 2. a new line is created before the new statement is created.
In both these cases, the default ending balance is 0.0 as the value is not set in the default_get method.
This PR fixes that and computes the ending balance accordingly
Case 1: Use the running balance of the line as the ending balance
Case 2: Add the line amount to the opening balance (since the line is not yet saved, the enterprise PR is required to set the amount in the context)