[FIX] website: make wysiwyg fonts independent from iframe fonts
Since the edited page is within an iframe, the built-in fonts used by
the wysiwyg are also loaded inside the page within the iframe, which
causes them to replace the one used when viewing the page as a visitor
if it happens to be one of the fallback fonts.

This commit avoids that by giving a distinct `font-family` to the fonts
used by the wysiwyg.

task-3080104