[FIX] hr_expense: not possible to update invoice_date
Before v16, posted expenses are of move_type 'entry'
After v16, of type 'in_receipt'.
But internally, we treat them as 'in_invoice' (bills).

So if a user created and posted an expense in v15, migrate to v16,
and want to update the move. He will get 'The Bill/Refund date
is required to validate this document.'

This commits allows the user to update the bill date on move of
type entry if there is an expense which is normally not possible.