[IMP] web: add a new property type: "separator"
Purpose
=======
In order to be able to create group of properties (like we can group
field in the views), we create a new property type, "separator".

Specification
=============
The separator create a group until the next separator, we can fold and
unfold the properties in a group.

The "fold" information is stored in the local storage, and is therefor
per user, but shared for all records of the same parent.

Task-3188915