[ADD] web_editor: add field html history
Create a new model mixins that allow other model to
activate the history feature on html fields.

This history automatically track the changes
on the related html field and allow the user
to revert to a previous version at any time.

This also introduce a new widget that can be used
in Odoo view to display the recent changes directly,
with option to revert to specific recent changes
directly in the UI.

task-3039787