[FIX] base: remove limit when read_group on expandable field
In both v14 and v15, loading a KanbanView triggers a web_read_group
without limit. That means that the KanbanView will display a column
by groupby_field value.

In v16, a limit of 80 has been introduced in
web/static/src/views/relational_models.js, DynamicList.DEFAULT_LIMIT = 80.
This means that loading a KanbanView triggers a web_read_group
with the limit parameter set to 80. But, since in KanbanView we usually
want to display empty columns, i.e. columns without any record
for the related groupby_field value, the groupby_field usually has
the group_expand attribute set to a callable.

Because web_read_group is also called with lazy=True, _read_group_fill_results
will be called at the end of _read_group_raw.
_read_group_fill_results will in turn expand the groupby results with
all the values returned by the field's group_expand callable. If there
are more possible values than limit=80, the result list will be filled
up by "no-record" dictionaries, i.e. dictionaries with
groupby_field__count set to 0.

On the js-side, the KanbanView limit is only used when calling
web_read_group. The JS does not truncate the results coming from the ORM.
So, when loading a KanbanView with more possible values of the
groupby_field than limit = 80, the KanbanView will show empty columns
without any records and that cannot be unfolded. This also impacts the ListView
when grouping by a group_expandable field.

There are probably multiple ways to fix that. Here the idea is to
drop the limit parameter (set to None) if the read_group is lazy and
if the first groupby_field has a group_expand callable set. That
way, the GROUP BY SQL query will return values for all the possible groups
and _read_group_fill_results will handle the displaying of empty columns.

opw-3173504